% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beads.R
\name{beads}
\alias{beads}
\title{Bead Plot}
\usage{
beads(predicted, reference)
}
\arguments{
\item{predicted}{a matrix of predicted class probabilities}

\item{reference}{a vector (factor) of classes to be used as the true result}
}
\description{
Bead Plot
}
