% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion Matrix}
\usage{
confusion_matrix(predicted, reference)
}
\arguments{
\item{predicted}{a factor of predicted class labels.}

\item{reference}{a factor of classes to be used as the true results.}
}
\description{
Graphical visualisation of a confusion matrix
}
