% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squares.R
\name{squares}
\alias{squares}
\title{Squares}
\usage{
squares(predicted, reference)
}
\arguments{
\item{predicted}{a vector of predicted class probabilities}

\item{reference}{a vector (factor) of classes to be used as the tru results}
}
\description{
Squares
}
